Initial implementation of the GitHub Action (#2)
# what

Initial implementation of the GitHub Action to load .env files and
export the key-value pair as environment variables with optional masking

# why

This PR was triggered by the fact that I liked [koheing/set-masked-env]
but it wasn't parsing the .env files correctly in all cases (e.g. it was
not stripping away comments). The original action was also using the old
version of Node.

I also thought I could use some JavaScript/TypeScript practice. 

# notes

This action can be used instead of [koheing/set-masked-env]. It adds an
optional `removeQuotes` parameter. Notice that the default value for
`mask` is `false` in this PR and `true` in [koheing/set-masked-env]

This action uses the
[parse-dotenv](https://github.com/luqmanoop/parse-dotenv) package for
parsing .env files

# todo

- [x] update README.md
- [ ] update tests

[koheing/set-masked-env]: https://github.com/koheing/set-masked-env

---------

Co-authored-by: João Roberto Fidellis <jrfidellis@gmail.com>